### 2019-10-16 #1814 @whedon
[PRE REVIEW]: MOAFS: A Massive Online Analysis library for featureselection in data streams
**Submitting author:** @mbdemoraes (<a href="http://orcid.org/0000-0002-9485-0334">Matheus Bernardelli de Moraes</a>)
**Repository:** <a href="https://github.com/mbdemoraes/moafs" target ="_blank">https://github.com/mbdemoraes/moafs</a>
**Version:** v1.0.0
