### 2019-10-18 #1814 @VivianePons
@mbdemoraes there seems to be a broken ref at the end of the paper, could you fix it? Thanks
