+++ paper.bib
@@ -31,10 +31,10 @@ address="Princeton, NJ",
 year=1946
 }
 
-@inproceedings{Carvalho2006,
+@article{Carvalho2006,
  author = {Carvalho, Vitor R. and Cohen, William W.},
  title = {Single-pass Online Learning: Performance, Voting Schemes and Online Feature Selection},
- booktitle = {Anais do 12 ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+ booktitle = {12 ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
  series = {KDD '06},
  year = {2006},
  isbn = {1-59593-339-5},
