### 2019-10-18 #1814 @mbdemoraes
Viviane,

Reference fixed! Thanks for the warning.

Att,

Matheus

Em sex, 18 de out de 2019 às 05:59, Viviane Pons <notifications@github.com>
escreveu:

> @mbdemoraes <https://github.com/mbdemoraes> there seems to be a broken
> ref at the end of the paper, could you fix it? Thanks
>
