### 2019-11-04 #1814 @mbdemoraes
Sure! From the joss reviewers list, I suggest:

   - jmadera
   - soroushj
   - xirdneh
   - dvysardana
   - nkratzke

Em seg., 4 de nov. de 2019 às 05:45, Viviane Pons <notifications@github.com>
escreveu:

> I am still looking for a second reviewer, @mbdemoraes
> <https://github.com/mbdemoraes> any suggestion?
>
