### 2019-11-21 #1814 @mbdemoraes
@VivianePons , given the trouble to find a suitable reviewer for this paper, I would like to check a possibility. There is someone, Andrés Bueno Barajas, who has a master's degree in Technology, focused on data streams and concept drift. He is experienced with the Massive Online Analysis framework. I have not suggested Andrés before because I've read JOSS Conflict of Interest and Andres has one published paper with André Gradvohl, a co-author of this paper. I would like to ask Andrés if he can give an impartial assessment of our work, but before I talk to him, I would like to know your opinion about it.
