### 2019-11-21 #1814 @VivianePons
That sounds like a good idea. I am ok with it if the reviewer himself feels he can do it. The JOSS policy does not forbid that, it only states that the conflict should be acknowledged so that the editor can decide. In this case, I think it is find especially as we also have a second reviewer.
