### 2019-11-21 #1814 @danielskatz
As a JOSS Associate-Editor-in-Chief, I think this depends on the details of the co-authored paper. For example: When was it published, and when was the work done? How many authors were there? Did they work closely together? Do they still work together?
