### 2019-11-21 #1814 @mbdemoraes
@danielskatz, the paper was published at the beginning of this year (2019) on a Brazilian journal, "Revista Brasileira de Computação Aplicada", but it refers to a work completed at the end of 2017. Andrés is a former student of University Campinas (not related to André Gradvohl), who helped a former André Gradvohl's thesis student, Luis F. Grim (the main author of the referred paper), in the practical approach of experiments, such as working with MATLAB, parallel libraries, and GPU. On the other hand, André Gradvohl, as an advisor of Luis F. Grim, helped him with theoretical aspects such as defining and validating the experiments, paper structuring, organization, and writing. Therefore, as far as I'm concerned, there was no direct collaboration between André Gradvohl and Andrés Barajas. And if they still work together, the answer is no, since André is a professor at the School of Technology at the University of Campinas and Andrés has concluded his master's degree in 2018 and is not at the university anymore.
