### 2019-11-21 #1814 @gradvohl
@danielskatz - Thank you for understanding. If you need more information, I am available to clarify and make the process as transparent as possible.
