### 2019-11-27 #1814 @gradvohl
Hello @VivianePons! Yes, we're trying to reach him by e-mail, but we haven't yet. We'll try one last time.
