### 2019-11-28 #1814 @mbdemoraes
Guys, we talked to Andrés and unfortunately he isn't available to review.
Other professionals we know have a strong conflict of interest and are not
suitable for this process. So we'll need your assistance with this task.

Em qua., 27 de nov. de 2019 às 12:23, André Leon S. Gradvohl <
notifications@github.com> escreveu:

> Hello @VivianePons <https://github.com/VivianePons>! Yes, we're trying to
> reach him by e-mail, but we haven't yet. We'll try one last time.
>
