### 2019-12-19 #1970 @whedon
[REVIEW]: MOAFS: A Massive Online Analysis library for featureselection in data streams
**Submitting author:** @mbdemoraes (<a href="http://orcid.org/0000-0002-9485-0334">Matheus Bernardelli de Moraes</a>)
**Repository:** <a href="https://github.com/mbdemoraes/moafs" target ="_blank">https://github.com/mbdemoraes/moafs</a>
**Version:** v1.1.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.11663307.v2" target="_blank">10.6084/m9.figshare.11663307.v2</a>

@sptennak, @DARSakthi, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
