### 2019-12-20 #1970 @arfon
@VivianePons - I just posted [a tweet](https://twitter.com/JOSS_TheOJ/status/1207837236454215682) on the JOSS account to see if we can find a second reviewer for this. I think we should have a second reviewer if at all possible.
