### 2019-12-20 #1970 @labarba
Hi all — I'm the EiC on rotation this week.
@sptennak — please go ahead with your review and post any questions here.
@VivianePons — we moved to a model of requiring two reviews a while back. Let us know if you need help adding a second reviewer here (it requires manually adding their checklist, after running `@wheddon add @user as reviewer`.
