### 2020-01-02 #1970 @sptennak
I started reviewing the paper.  At this point, I could locate the repository, install the software and run the GUI. The software installed and GUI loads as specified in the documentation. Good work from the authors!  However, I see some room for improvements there. I greatly appreciate if the author can provide some details on using the software to allow me testing the functionality.

I also suggest authors to
[1] add some illustrations/text content to the paper describing the concepts and design of the software in very high level. That will help the reader to better understand the software.
[2] provide some instructions (or link) in the documentation on how to install MOA. The link appear on README.md in github, but not in the documentation.
[3] provide some sample output/results for the examples mentioned in the documentation.
[4] provide details on components to GUI and how a user can navigate through GUI using different examples.
[5] provide sample arffFile to test the example in the documentation.

Thanks
