### 2020-01-06 30bc06a
Sample outputs and detailed GUI documentation
