+++ docs/examplesGUI.rst
@@ -19,7 +19,13 @@ If everything is OK, MOA's GUI should appear as illustrated in the figure below.
 .. image:: _static/moafs_01.png
   :width: 100 %
 
-Then, click on ``Configure`` button on the left side. The ``Configure Task`` window should appear. Select ``class moa.tasks.EvaluateInterleavedTestThenTrain`` in the
+Classification without feature selection (No method)
+----------------------------------------------------
+
+.. image:: _static/moafs_01.png
+  :width: 100 %
+
+Click on ``Configure`` button on the left side. The ``Configure Task`` window should appear. Select ``class moa.tasks.EvaluateInterleavedTestThenTrain`` in the
 first dropdown list.
 
 .. image:: _static/moafs_02.png
@@ -27,7 +33,8 @@ first dropdown list.
 
 Still on the ``Configure Task`` window, on ``learner`` options, click on the ``Edit`` button. The ``Editing option: learner`` window should be presented. Select
 ``class moa.featureselection.classifiers.NaiveBayes`` on the first dropdown list. Three options must be presented as ilustrated in the figure below. 
-With these three options, you can select the number of features to be selected, the feature selection method and the processing window size.
+With these three options, you can select the number of features to be selected, the feature selection method and the processing window size. To perform
+classification without feature selection, set ``fsMethods`' to 0.
 
 .. image:: _static/moafs_03.png
   :width: 50 %
@@ -38,50 +45,16 @@ To select a data set from a local directory, on the ``Configure Task`` window, o
   :width: 50 %
 
 Then, ``Editing option: stream`` window should be presented. Select
-``class moa.streams.ArffFileStream`` on the first dropdown list. Then you can select the data set from the ``arffFile`` option.
+``class moa.streams.ArffFileStream`` on the first dropdown list. Then you can select the data set from the ``arffFile`` option. Click ``OK`` and then ``OK`` again.
 
 .. image:: _static/moafs_04.png
   :width: 50 %
 
-If everything is according to plan, just press the ``OK`` button on all windows and you will be returned to the main page. There, just click on the ``Run`` button on the right side and MOA
+If everything is according to plan, just press the ``OK`` button on all windows and you will be returned to the main window. There, just click on the ``Run`` button on the right side and MOA
 will perform the classification of the data set using the selected feature selection method.
 
 .. image:: _static/moafs_05.png
   :width: 100 %
-
-
-Classification without feature selection (No method)
-----------------------------------------------------
-
-.. image:: _static/moafs_01.png
-  :width: 100 %
-
-.. image:: _static/moafs_02.png
-  :width: 50 %
-
-.. image:: _static/moafs_03.png
-  :width: 50 %
-
-Select a desired number of features on ``numFeatures`` and the window size on ``winSize`` options. For the ``fsMethod``, leave it as zero. Therefore, no feature
-selection method will be used for the classification process. 
-
-To select a data set from a local directory, on the ``Configure Task`` window, on ``stream`` options, click on the ``Edit`` button. 
-
-.. image:: _static/moafs_02.png
-  :width: 50 %
-
-Then, ``Editing option: stream`` window should be presented. Select
-``class moa.streams.ArffFileStream`` on the first dropdown list. Then you can select the data set from the ``arffFile`` option.
-
-.. image:: _static/moafs_04.png
-  :width: 50 %
-
[…]
