+++ sample outputs/sample_no_method_output.txt
@@ -1,0 +1,16 @@
+learning evaluation instances,evaluation time (cpu seconds),model cost (RAM-Hours),classified instances,classifications correct (percent),Kappa Statistic (percent),Kappa Temporal Statistic (percent),Kappa M Statistic (percent),F1 Score (percent),F1 Score for class 0 (percent),F1 Score for class 1 (percent),Precision (percent),Precision for class 0 (percent),Precision for class 1 (percent),Recall (percent),Recall for class 0 (percent),Recall for class 1 (percent),model training instances,model serialized size (bytes)
+100.0,0.024279652,4.153105198095242E-10,100.0,52.0,19.81289675910457,27.272727272727277,-45.45454545454547,65.56539045549542,45.45454545454545,57.14285714285714,67.87221217600964,95.23809523809523,40.50632911392405,63.410221619176845,29.850746268656714,96.96969696969697,100.0,66120.0
+200.0,0.053097501,9.093217697553337E-10,200.0,56.00000000000001,25.10000851136267,33.83458646616542,-31.343283582089548,68.46873313338206,51.11111111111111,60.0,70.50479766374636,97.87234042553192,43.13725490196079,66.54696442599035,34.58646616541353,98.50746268656717,200.0,66264.0
+300.0,0.069077784,1.1833967687934638E-9,300.0,70.0,44.44444444444444,54.99999999999999,9.999999999999996,76.22573856107257,71.3375796178344,68.53146853146853,75.46689303904924,98.24561403508771,52.68817204301075,77.0,56.00000000000001,98.0,300.0,66296.0
+400.0,0.084274511,1.4451339310283463E-9,400.0,57.49999999999999,15.070067194564482,36.32958801498127,-1.7964071856287573,57.64667970592677,60.64814814814815,53.80434782608695,57.541438535963394,65.82914572864321,49.25373134328358,57.752306545706865,56.22317596566524,59.2814371257485,400.0,66576.0
+500.0,0.094687591,1.62467506164064E-9,500.0,52.800000000000004,4.692674258945165,29.341317365269465,-1.2875536480686707,52.35262249259676,57.24637681159419,47.32142857142858,52.37046103631171,55.43859649122807,49.30232558139535,52.33479609715324,59.176029962546814,45.493562231759654,500.0,66648.0
+600.0,0.104967688,1.8020296933626136E-9,600.0,53.666666666666664,7.333333333333325,30.67331670822942,7.333333333333325,53.69982866612365,56.56250000000001,50.357142857142854,53.73303167420814,53.23529411764706,54.230769230769226,53.666666666666664,60.333333333333336,47.0,600.0,66688.0
+700.0,0.114019589,1.9582514027609593E-9,700.0,56.00000000000001,11.372736017363717,34.1880341880342,7.507507507507527,55.713986274724405,59.89583333333333,51.26582278481013,55.768605242745984,57.356608478803,54.18060200668896,55.659474188084545,62.67029972752044,48.64864864864865,700.0,66712.0
+800.0,0.122997526,2.1131966087470452E-9,800.0,60.12499999999999,19.216987224128538,40.37383177570093,13.07901907356947,59.63475897500621,64.35754189944134,54.751773049645394,59.71912702681934,62.33766233766234,57.100591715976336,59.55062896841628,66.51270207852194,52.58855585831063,800.0,66712.0
[…]
