### 2020-01-06 46073ea
Corrections on readme
