### 2020-01-06 b993a3b
Corrections on readme
