### 2020-01-06 5a3b4e6
Corrections on readme
