### 2020-01-06 f48ace8
Corrections on readme
