### 2020-01-06 98484a4
Adding ilustrations and text content on paper
