### 2020-01-06 df6fb46
Changing image extension
