### 2020-01-06 4a2001d
Fixing documentation typos
