### 2020-01-06 #1970 @mbdemoraes
@sptennak I added all your suggestions. Can you please verify it? Concerning how to test the software, I included a section in both readme and documentation about it. As you'll see, it is rather simple, just download a dataset in arff format which you want to experiment, run MOA with MOAFS and set the parameters accordingly to what you desire. Then run it. I think it will be easier now that I've included examples using GUI for each method.  Please verify if it is clear and easy to understand now. If not, let me know.

@VivianePons just a question: I've made changes in the paper, so I think it needs to be generated again. It's me, you or the reviewers who run the "whedon generate pdf" command?

Thank you!
