### 2020-01-15 #1970 @DARSakthi
The authors have presented a package which makes the analysis of high throughput data more accurate, while also making handling this analysis easier.

Some minor comments on the paper:
     - I suggest you switch the content in Section: Summary with Section: Statement of Need and make necessary changes for flow and comprehension. The reason I recommend this is because what has been written in each section fulfils the requirement of the other section. In Summary you have told us: what is the challenge the software provides a solution to. In Statement of Need you summarise for us how the software works.
     - You should say somewhere telling us explicitly what MOAFS stands for.
     - There should be a line or two somewhere telling us in concise, explicit terms what MOAFS adds to MOA. Currently the section in which you summarise the package has a long explanation, but it is missing a simple, one line statement about what exactly it does. Preferably it will be placed at the beginning or end of the paragraph. A good item to use would be an abridged version of the abstract contained in moafs/README.md
     - It would be wise to clarify the exact role of Naive Bayes here, or else avoid talking about it at all. Statement of Need makes it seem like Naive Bayes classifier suggests a feature selection algorithm to use. As I understand it, MOAFS is a sandbox for the user to test out the best FS algorithm and Naive Bayes comes later.  You should emphasise this.
     - While you are editing this content I suggest you make edits to the language in the current summary section. There, language and sentence structure could use some revision.

The package itself is well done and functions as described. The extensions are well integrated, showing a good understanding of MOA's original backed, and the code performs efficiently.
     - In documentation (see: https://moafs.readthedocs.io/en/latest/install.html#running-gui) I would recommend making clear that the scripts needs to refer to the full path for sizeof.jar. Here's an example: "[...rest of script...] full_path_to_folder\lib\sizeofag-1.0.4.jar"
     - Alternatively, specify that this must take place from the lib directory.
The above is not strictly necessary, but a suggestion for user clarity.

I recommend eventually bringing the package up to date with MOA 2019 versions, but this can be a living repository so to speak, with that goal coming later. When this happens be sure to update the paper and docs.

I have not tried it on an independent data set but I do not see that as necessary. If I have been unclear on anything or you seek further suggestions please respond. Well done -- I look forward to seeing the final iteration of the paper.
