### 2020-01-17 b6b50fd
Applying suggestions from reviewers
