### 2020-01-17 f090a1e
Applying suggestions from reviewers
