### 2020-01-17 #1970 @mbdemoraes
@DARSakthi  Your suggestions were applied to both paper and documentation. Here is the list of modifications:

- Switched contents from Summary and Statement of Need sections;
- Included a sentence explaining what MOAFS stands for;
- Included a concise, one-line statement about MOAFS purpose in the first paragraph (Summary);
- Changed the "Library design" paragraph to explain better that feature selection is optional and the classification may occur without it;
- Revised paper language and sentence structure;
- On documentation, specified that scripts may be executed from the lib folder or using the full path.

Concerning the update to the 2019 version, you are right, I intend to make a living repository and upgrade it later.

Please, confirm if the adjustments meet your suggestions!
