### 2020-01-18 bc61a10
Including docs and license markdowns
