### 2020-01-20 #1970 @VivianePons
Perfect, so I understand we have a go from both reviewers! Let's move forward then to the final steps of accepting the paper :) :)
