### 2020-01-20 #1970 @VivianePons
Hi @mbdemoraes can you please check the paper one last time and see if everything is alright? Especially, see if you want to do something about @DARSakthi latest minor comment.

Once done, could you make an archive (on Zenodo, figshare or what you want) of the current repo and send me the DOI. Thank you!
