### 2020-01-21 #1970 @labarba
@mbdemoraes 👋 — I've updated the version according to what I see on the repo. Please confirm that is what you wanted. Also, please update the metadata of your Figshare deposit so that the title of the archive matches the JOSS paper, and the list of authors also matches. (It would be nice if you added a description, too.)
