### 2020-01-21 #1970 @mbdemoraes
@labarba Thank you for updating the version, it is correct now.

I also updated the title and description of our Figshare deposit. Concerning the list of authors, it does already matches, doesn't it? The only difference is that Figshare uses an abbreviated form. I had to use André Gradvohl's ORCID to find him at Figshare, and since he registered his name in an abbreviated form, Fighsare used that abbreviation. If this is a problem, please inform me, because I will have to update the paper.
