+++ README.md
@@ -1,7 +1,6 @@
 
 [![docs](https://readthedocs.org/projects/moafs/badge/?version=latest)](https://moafs.readthedocs.io/en/latest/?badge=latest)
-[![license](https://img.shields.io/eclipse-marketplace/l/Vrapper)](
-https://github.com/mbdemoraes/moafs/blob/master/LICENSE)
+[![license](https://img.shields.io/eclipse-marketplace/l/Vrapper)](https://github.com/mbdemoraes/moafs/blob/master/LICENSE)
 [![JOSS](https://joss.theoj.org/papers/10.21105/joss.01970/status.svg)](https://doi.org/10.21105/joss.01970)
 
 
